MOVE: Reorganize environment modules
Relocated `scenario_type` and `terminal_access_tool_provider` modules to the `environment` package. Updated import statements across multiple files to reflect the new structure. This improves code organization and module categorization.
