+++ src/ideformer_client/environment/scenario_environment_manager.py
@@ -4,8 +4,8 @@ from docker.models.containers import Container
 
 import logging
 
-from src.ideformer_client.exceptions import ScenarioEnvironmentException
-from src.ideformer_client.scenario_type import ScenarioType
+from src.ideformer_client.utils.exceptions import ScenarioEnvironmentException
+from src.ideformer_client.environment.scenario_type import ScenarioType
 from src.yt_scripts.schemas import RepositoryDataRow
 
 class ScenarioEnvironmentManager:
